loans2grow/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.4
greenlet==3.0.3
mpmath==1.3.0
multipledispatch==1.0.0
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==3.10.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.3
SQLAlchemy==2.0.21
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
